Cleanup, TypeScript refactoring.
