+++ huex1/src/tools/tool/brush.ts
@@ -1,10 +1,12 @@
 import paper from 'paper';
 import store from '../../store';
-import {createLayer} from '../shared';
-import history from '../history';
-import {DrawAction} from "../action";
+import {createLayer, hexToRgba} from '../shared';
 
-let local = {
+interface iBrush {
+  path: paper.Path | null
+}
+
+let local: iBrush = {
   path: null
 }
 
@@ -13,12 +15,12 @@ const toolProperties = store.state.toolProperties.brush
 function onMouseDown() {
   let layer = createLayer();
   local.path = new paper.Path();
-  const rgba = hexToRgba(toolProperties.color || '#000000ff');
-  local.path.fillColor = `rgba(${rgba.r},${rgba.g},${rgba.b},${rgba.a / 255})`;
+  const rgba = hexToRgba(toolProperties.color);
+  local.path.fillColor = `rgba(${rgba.r},${rgba.g},${rgba.b},${rgba.a / 255})` as unknown as paper.Color;
   layer.addChild(local.path);
 }
 
-function onMouseDrag(event) {
+function onMouseDrag(event: paper.ToolEvent) {
   if (!local.path) return;
   const step = event.delta;
   step.x *= toolProperties.size / 10;
@@ -33,18 +35,9 @@ function onMouseDrag(event) {
 
 
 function onMouseUp() {
+  if (!local.path) return
+
   local.path.simplify(1);
-  const action = new DrawAction({
-    layer: local.path.layer.name,
-    tool: store.getters.activeTool,
-    points: local.path.segments.map(seg => {
-      return {
-        x: seg._point._x,
-        y: seg._point._y
-      }
-    })
-  });
-  history.add(action);
   local.path.selected = false;
 
   const json = local.path.layer.exportJSON()
@@ -59,13 +52,3 @@ tool.maxDistance = 7;
 tool.onMouseDown = onMouseDown;
 tool.onMouseDrag = onMouseDrag;
 tool.onMouseUp = onMouseUp;
-
-function hexToRgba(hex) {
-  let result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
-  return result ? {
-    r: parseInt(result[1], 16),
-    g: parseInt(result[2], 16),
-    b: parseInt(result[3], 16),
-    a: parseInt(result[4], 16)
-  } : null;
-}
