--- huex1/src/tools/tool/tools.js
@@ -1,13 +1,0 @@
-import {tool as brush} from './brush';
-import {tool as eraser} from './eraser';
-import {tool as square} from './square';
-import {tool as triangle } from './triangle';
-import { tool as circle } from './circle';
-
-export default {
-  brush,
-  eraser,
-  square,
-  circle,
-  triangle
-}
