+++ huex1/src/tools/tool/tools.ts
@@ -1,0 +1,13 @@
+import { tool as brush } from './brush';
+import { tool as eraser } from './eraser';
+import { tool as square } from './rectangle';
+import { tool as triangle } from './triangle';
+import { tool as circle } from './circle';
+
+export default {
+  brush,
+  eraser,
+  square,
+  circle,
+  triangle
+}
