+++ huex1/src/tools/tool/triangle.ts
@@ -1,10 +1,14 @@
 import paper from 'paper';
 import store from '../../store';
-import history from '../history';
-import {createLayer} from '../shared';
-import {DrawAction} from '../action';
+import {createLayer, hexToRgba} from '../shared';
 
-let local = {
+interface iTriangle {
+  path: paper.Path | null,
+  center: paper.Point | null,
+  layer: paper.Layer | null
+}
+
+let local: iTriangle = {
   path: null,
   center: null,
   layer: null
@@ -12,38 +16,30 @@ let local = {
 
 const toolProperties = store.state.toolProperties.triangle
 
-function onMouseDown(event) {
+function onMouseDown(event: paper.ToolEvent) {
   local.layer = createLayer();
   local.center = event.point;
 }
 
-function onMouseDrag(event) {
+function onMouseDrag(event: paper.ToolEvent) {
+  if(!local.center) return
   if (local.path) {
     local.path.remove();
   }
+
   local.path = new paper.Path.RegularPolygon(local.center, 3, Math.sqrt((event.point.x - local.center.x) * (event.point.x - local.center.x) + (event.point.y - local.center.y) * (event.point.y - local.center.y)));
   const rgba = hexToRgba(toolProperties.color);
-  local.path.strokeColor = `rgba(${rgba.r},${rgba.g},${rgba.b},${rgba.a / 255})`;
+  local.path.strokeColor = `rgba(${rgba.r},${rgba.g},${rgba.b},${rgba.a / 255})` as unknown as paper.Color;
   local.path.strokeWidth = toolProperties.size;
 }
 
 function onMouseUp() {
+  if (!local.layer || !local.path) return
   local.layer.addChild(local.path);
-  const action = new DrawAction({
-    layer: local.path.layer.name,
-    tool: store.getters.activeTool,
-    points: local.path.segments.map(seg => {
-      return {
-        x: seg._point._x,
-        y: seg._point._y
-      }
-    })
-  });
 
   const json = local.path.layer.exportJSON()
   store.dispatch('addElementToCurrentBoardContent', json)
 
-  history.add(action);
   local.path = null;
 }
 
@@ -51,13 +47,3 @@ export const tool = new paper.Tool();
 tool.onMouseDown = onMouseDown;
 tool.onMouseDrag = onMouseDrag;
 tool.onMouseUp = onMouseUp;
-
-function hexToRgba(hex) {
-  let result = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
-  return result ? {
-    r: parseInt(result[1], 16),
-    g: parseInt(result[2], 16),
-    b: parseInt(result[3], 16),
-    a: parseInt(result[4], 16)
-  } : null;
-}
